requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
celery
redis
httpx